app/requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
spotipy==2.23.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.25.2
scikit-learn==1.3.0
scipy==1.11.1